Problem: extensive use of macros in supabase-wrappers
There are at least two problems with this:

1. Development experience

Most of the code is in macros and becomes hard to reason about and
navigate using IDEs as it is all simply quoted expressions.

2. One macro approach captures all provided FDWs and will conflict

"Magic" macro provides pg_extern function that exposes all listed FDWs
in one bundle. However, should any other extension use
supabase-wrappers, it won't install alongside `wrappers` extension
because of a duplicate symbol.

3. Compile-times may suffer with macro-heavy code.

Solution: move all this code to supabase-wrappers

As my analysis shown, there's no real need for macros. In fact, in some
parts (like use of ForeignDataWrapper::new without it being part of the
trait) are downright wrong.

For the time being, I've removed the "meta" listing functionality as it
would need to be re-designed to support multiple extensions and/or
unbundled packaging of multiple FDWs. If it will be deemed necessary,
I can see a number of ways it can re-implemented.

There's a lot that can be improved, but I've tried to keep this change
short and to the point (and that point is removing the dependency on
macros)
