--- supabase-wrappers-macros/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "supabase-wrappers-macros"
-version = "0.1.4"
-edition = "2021"
-authors = ["Supabase Inc. https://supabase.com/"]
-license = "Apache-2.0"
-description = "Postgres Foreign Data Wrapper development framework macros for supabase-wrappers"
-homepage = "https://github.com/supabase/wrappers/tree/main/supabase-wrappers-macros"
-repository = "https://github.com/supabase/wrappers/tree/main/supabase-wrappers-macros"
-categories = ["database"]
-keywords = ["database", "postgres", "postgresql", "extension"]
-
-[dependencies]
-syn = { version = "1.0", features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
-
-[lib]
-proc-macro = true
-
