--- supabase-wrappers-macros/rustfmt.toml
@@ -1,1 +1,0 @@
-max_width = 100
