--- supabase-wrappers-macros/src/parser/mod.rs
@@ -1,3 +1,0 @@
-mod fdw_type;
-
-pub(crate) use fdw_type::FdwType;
