+++ wrappers/Cargo.toml
@@ -24,6 +24,9 @@ firebase_fdw = ["reqwest", "reqwest-middleware", "reqwest-retry", "serde_json", 
 # TODO: audit dependencies
 airtable_fdw = ["reqwest", "reqwest-middleware", "reqwest-retry", "serde_json", "serde", "url"]
 
+# Does not include helloworld_fdw because of its general uselessness
+all_fdws = ["airtable_fdw", "bigquery_fdw", "clickhouse_fdw", "stripe_fdw", "firebase_fdw"]
+
 [dependencies]
 pgx = { version = "=0.6.0", features = ["time-crate"] }
 cfg-if = "1.0"
